# backend/cmd/docs/openapi.yaml
openapi: 3.1.0
info:
  title: API Приложения для Аутентификации и Опросов
  description: API для управления опросами, вопросами, ответами, аутентификацией пользователей и аналитикой.
  version: 1.0.0
servers:
  - url: /api
    description: Основной сервер API
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке
        details:
          type: string
          description: Детали ошибки (опционально)
      required:
        - error
    Survey:
      type: object
      properties:
        id:
          type: integer
          description: Идентификатор опроса
        title:
          type: string
          description: Название опроса
        description:
          type: string
          description: Описание опроса
        created_by:
          type: integer
          description: ID пользователя, создавшего опрос
        created_at:
          type: string
          format: date-time
          description: Дата создания
        max_ball:
          type: integer
          description: Максимальный балл за опрос
      required:
        - title
    Question:
      type: object
      properties:
        id:
          type: integer
          description: Идентификатор вопроса
        survey_id:
          type: integer
          description: ID опроса
        question_text:
          type: string
          description: Текст вопроса
        is_required:
          type: boolean
          description: Обязателен ли вопрос
        is_test:
          type: boolean
          description: Является ли вопрос тестовым
        ball:
          type: integer
          description: Баллы за вопрос
        correct_answer:
          type: string
          description: Правильный ответ (для текстовых вопросов)
        answers:
          type: array
          items:
            $ref: '#/components/schemas/AnswerOption'
          description: Варианты ответов (для тестовых вопросов)
      required:
        - survey_id
        - question_text
    AnswerOption:
      type: object
      properties:
        id:
          type: integer
          description: Идентификатор варианта ответа
        answer_text:
          type: string
          description: Текст ответа
        correct:
          type: boolean
          description: Является ли ответ правильным
      required:
        - answer_text
...
    SurveyAnalytics:
      type: object
      properties:
        survey:
          $ref: '#/components/schemas/Survey'
          description: Информация об опросе
        total_responses:
          type: integer
          description: Общее количество ответов
        average_score:
          type: number
          format: float
          description: Средний балл
        participants:
          type: array
          items:
            type: object
            properties:
              username:
                type: string
                description: Имя пользователя
              total_ball:
                type: integer
                description: Общий балл
              date:
                type: string
                format: date-time
                description: Дата ответа
          description: Участники опроса
        questions:
          type: array
          items:
            type: object
            properties:
              question:
                $ref: '#/components/schemas/Question'
                description: Вопрос
              correct_rate:
                type: number
                format: float
                description: Процент правильных ответов
              answer_stats:
                type: array
                items:
                  type: object
                  properties:
                    answer_text:
                      type: string
                      description: Текст ответа
                    count:
                      type: integer
                      description: Количество выборов
                    is_correct:
                      type: boolean
                      description: Правильный ли ответ
                description: Статистика по вариантам ответа
              text_responses:
                type: array
                items:
                  type: object
                  properties:
                    username:
                      type: string
                      description: Имя пользователя
                    answer:
                      type: string
                      description: Текстовый ответ
                description: Текстовые ответы
          description: Статистика по вопросам
paths:
...
  /surveys:
    get:
      summary: Получение списка всех опросов
      description: Возвращает список всех доступных опросов.
      responses:
        '200':
          description: Список опросов успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Survey'
        '500':
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /surveys/{id}/analytics:
    get:
      summary: Получение аналитики опроса
      description: Возвращает аналитику по указанному опросу (для создателя).
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID опроса
      responses:
        '200':
          description: Аналитика успешно получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveyAnalytics'
        '400':
          description: Неверный ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Требуется аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Опрос не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /surveys/{id}/analytics/pdf:
    get:
      summary: Получение аналитики опроса в формате PDF
      description: Возвращает аналитику по указанному опросу в формате PDF (для создателя).
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID опроса
      responses:
        '200':
          description: PDF успешно сгенерирован
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Неверный ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Требуется аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Опрос не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...